Rename cmd.py -> ncmd.py because it conflicted with Python's cmd module and caused issues "explicit is better than implicit" my &ss
