+++ python/scope.py
@@ -13,7 +13,7 @@ from type import NType, NGenericType, NAliasType, NTypeVars, NModule, apply_gene
 from enums import EnumType, EnumValue, EnumPattern
 from native_function import NativeFunction
 from native_types import n_list_type, n_cmd_type
-from cmd import Cmd
+from ncmd import Cmd
 from type_check_error import TypeCheckError, display_type
 from display import display_value
 from operation_types import binary_operation_types, unary_operation_types, comparable_types, iterable_types, legacy_iterable_types
